+++ demos/employees/src/main/java/is/codion/demos/employees/testing/scenarios/InsertEmployee.java
@@ -31,7 +31,7 @@ public final class InsertEmployee extends AbstractPerformer {
 
 	@Override
 	public void perform(EmployeesAppModel application) throws EntityValidationException {
-		SwingEntityModel departmentModel = application.entityModels().get(Department.TYPE);
+		SwingEntityModel departmentModel = application.models().get(Department.TYPE);
 		selectRandomRow(departmentModel.tableModel());
 		SwingEntityModel employeeModel = departmentModel.detail().get(Employee.TYPE);
 		Entity employee = new DefaultEntityFactory(application.connection()).entity(Employee.TYPE);