+++ demos/employees/src/main/java/is/codion/demos/employees/testing/scenarios/SelectDepartment.java
@@ -26,7 +26,7 @@ public final class SelectDepartment extends AbstractPerformer {
 
 	@Override
 	public void perform(EmployeesAppModel application) {
-		selectRandomRow(application.entityModels().get(Department.TYPE).tableModel());
+		selectRandomRow(application.models().get(Department.TYPE).tableModel());
 	}
 }
 // end::loadTest[]